Pandas Excel.py File
So, basically i was trying to read and write excel file using pandas. So it was able to read both .xlsx and .XLSX files but while writing to another file with extension .XLS it was erroring out so i have changed execl.py file a little bit. Now it can save file with both .xlsx and .XLSX extension.